requirements.txt
numpy==1.17.3
selenium==3.141.0
PyMySQL==0.9.3
urllib3==1.24.3
website==1.1
xlrd==1.2.0
opencv_python==4.4.0.42
requests==2.24.0
xpinyin==0.5.6
XlsxWriter==1.3.3
Appium_Python_Client==1.0.2
ddt==1.4.1
Pillow==8.0.1
requests_toolbelt==0.9.1
pyopenssl==19.1.0
jsonschema==3.2.0